Closes #175 - Added newTransaction() method to Fluo API
