+++ modules/core/src/main/java/io/fluo/core/exceptions/AlreadyAcknowledgedException.java
@@ -15,6 +15,8 @@
  */
 package io.fluo.core.exceptions;
 
+import io.fluo.api.exceptions.CommitException;
+
 /**
  * 
  */
