+++ modules/mapreduce/src/main/java/io/fluo/mapreduce/FluoInputFormat.java
@@ -22,9 +22,6 @@ import java.util.List;
 import java.util.Map.Entry;
 import java.util.Properties;
 
-import org.apache.commons.configuration.ConfigurationConverter;
-
-import org.apache.commons.configuration.PropertiesConfiguration;
 import io.fluo.api.config.FluoConfiguration;
 import io.fluo.api.config.ScannerConfiguration;
 import io.fluo.api.data.Bytes;
@@ -33,11 +30,12 @@ import io.fluo.api.iterator.ColumnIterator;
 import io.fluo.api.iterator.RowIterator;
 import io.fluo.core.impl.Environment;
 import io.fluo.core.impl.TransactionImpl;
-import io.fluo.core.oracle.OracleClient;
 import io.fluo.core.util.SpanUtil;
 import org.apache.accumulo.core.client.mapreduce.AccumuloInputFormat;
 import org.apache.accumulo.core.client.mapreduce.RangeInputSplit;
 import org.apache.accumulo.core.client.security.tokens.PasswordToken;
+import org.apache.commons.configuration.ConfigurationConverter;
+import org.apache.commons.configuration.PropertiesConfiguration;
 import org.apache.hadoop.mapreduce.InputFormat;
 import org.apache.hadoop.mapreduce.InputSplit;
 import org.apache.hadoop.mapreduce.Job;
